Extract helm download code to public package
This allows code re-use with other systems and projects.
