+++ cmd/apps/istio_app.go
@@ -10,7 +10,7 @@ import (
 	"github.com/alexellis/k3sup/pkg"
 	"github.com/alexellis/k3sup/pkg/config"
 	"github.com/alexellis/k3sup/pkg/env"
-
+	"github.com/alexellis/k3sup/pkg/helm"
 	"github.com/spf13/cobra"
 )
 
@@ -66,7 +66,7 @@ func MakeInstallIstio() *cobra.Command {
 			os.Setenv("HELM_VERSION", helm3Version)
 		}
 
-		_, err = tryDownloadHelm(userPath, clientArch, clientOS, helm3)
+		_, err = helm.TryDownloadHelm(userPath, clientArch, clientOS, helm3)
 		if err != nil {
 			return err
 		}
