+++ cmd/apps/kubernetes_exec.go
@@ -2,18 +2,14 @@ package apps
 
 import (
 	"fmt"
-	"io/ioutil"
 	"log"
-	"net/http"
-	"net/url"
 	"os"
 	"path"
 	"strings"
 
 	execute "github.com/alexellis/go-execute/pkg/v1"
+	"github.com/alexellis/k3sup/pkg/env"
 )
-
-const helmVersion = "v2.16.0"
 
 func fetchChart(path, chart string, helm3 bool) error {
 
@@ -29,7 +25,7 @@ func fetchChart(path, chart string, helm3 bool) error {
 	}
 
 	task := execute.ExecTask{
-		Command:     fmt.Sprintf("%s fetch %s --untar --untardir %s", localBinary("helm", subdir), chart, path),
+		Command:     fmt.Sprintf("%s fetch %s --untar --untardir %s", env.LocalBinary("helm", subdir), chart, path),
 		Env:         os.Environ(),
 		StreamStdio: true,
 	}
@@ -79,7 +75,7 @@ func helm3Upgrade(basePath, chart, namespace, values string, overrides map[strin
 	}
 
 	task := execute.ExecTask{
-		Command:     localBinary("helm", "helm3"),
+		Command:     env.LocalBinary("helm", "helm3"),
 		Args:        args,
 		Env:         os.Environ(),
 		Cwd:         basePath,
@@ -131,7 +127,7 @@ func templateChart(basePath, chart, namespace, outputPath, values string, overri
 
 	task := execute.ExecTask{
 		Command: fmt.Sprintf("%s template %s --name %s --namespace %s --output-dir %s %s %s",
-			localBinary("helm", ""), chart, chart, namespace, outputPath, valuesStr, overridesStr),
+			env.LocalBinary("helm", ""), chart, chart, namespace, outputPath, valuesStr, overridesStr),
 		Env:         os.Environ(),
 		Cwd:         basePath,
 		StreamStdio: true,
@@ -154,16 +150,6 @@ func templateChart(basePath, chart, namespace, outputPath, values string, overri
 	return nil
 }
 
-func localBinary(name, subdir string) string {
-	home := os.Getenv("HOME")
-	val := path.Join(home, ".k3sup/bin/")
-	if len(subdir) > 0 {
-		val = path.Join(val, subdir)
-	}
-
-	return path.Join(val, name)
-}
-
 func addHelmRepo(name, url string, helm3 bool) error {
 	subdir := ""
 	if helm3 {
@@ -171,7 +157,7 @@ func addHelmRepo(name, url string, helm3 bool) error {
 	}
 
 	task := execute.ExecTask{
-		Command:     fmt.Sprintf("%s repo add %s %s", localBinary("helm", subdir), name, url),
+		Command:     fmt.Sprintf("%s repo add %s %s", env.LocalBinary("helm", subdir), name, url),
 		Env:         os.Environ(),
 		StreamStdio: true,
 	}
@@ -193,31 +179,8 @@ func updateHelmRepos(helm3 bool) error {
 		subdir = "helm3"
 	}
 	task := execute.ExecTask{
-		Command:     fmt.Sprintf("%s repo update", localBinary("helm", subdir)),
+		Command:     fmt.Sprintf("%s repo update", env.LocalBinary("helm", subdir)),
 		Env:         os.Environ(),
-		StreamStdio: true,
-	}
-
-	res, err := task.Execute()
-
-	if err != nil {
-		return err
-	}
-
-	if res.ExitCode != 0 {
-		return fmt.Errorf("exit code %d", res.ExitCode)
-	}
-	return nil
-}
-
-func helmInit() error {
-	fmt.Printf("Running helm init.\n")
-	subdir := ""
-
-	task := execute.ExecTask{
-		Command:     fmt.Sprintf("%s", localBinary("helm", subdir)),
-		Env:         os.Environ(),
-		Args:        []string{"init", "--client-only"},
 		StreamStdio: true,
 	}
 
@@ -274,69 +237,4 @@ func getDefaultKubeconfig() string {
 	}
 
 	return kubeConfigPath
-}
-
-func tryDownloadHelm(userPath, clientArch, clientOS string, helm3 bool) (string, error) {
-	helmVal := "helm"
-	if helm3 {
-		helmVal = "helm3"
-	}
-
-	helmBinaryPath := path.Join(path.Join(userPath, "bin"), helmVal)
-	if _, statErr := os.Stat(helmBinaryPath); statErr != nil {
-		subdir := ""
-		if helm3 {
-			subdir = "helm3"
-		}
-		downloadHelm(userPath, clientArch, clientOS, subdir)
-
-		if !helm3 {
-			err := helmInit()
-			if err != nil {
-				return "", err
-			}
-		}
-	}
-	return helmBinaryPath, nil
-}
-
-func getHelmURL(arch, os, version string) string {
-	archSuffix := "amd64"
-	osSuffix := strings.ToLower(os)
-
-	if strings.HasPrefix(arch, "armv7") {
-		archSuffix = "arm"
-	} else if strings.HasPrefix(arch, "aarch64") {
-		archSuffix = "arm64"
-	}
-
-	return fmt.Sprintf("https://get.helm.sh/helm-%s-%s-%s.tar.gz", version, osSuffix, archSuffix)
-}
-
-func downloadHelm(userPath, clientArch, clientOS, subdir string) error {
-	useHelmVersion := helmVersion
-	if val, ok := os.LookupEnv("HELM_VERSION"); ok && len(val) > 0 {
-		useHelmVersion = val
-	}
-
-	helmURL := getHelmURL(clientArch, clientOS, useHelmVersion)
-	fmt.Println(helmURL)
-	parsedURL, _ := url.Parse(helmURL)
-
-	res, err := http.DefaultClient.Get(parsedURL.String())
-	if err != nil {
-		return err
-	}
-
-	dest := path.Join(path.Join(userPath, "bin"), subdir)
-	os.MkdirAll(dest, 0700)
-
-	defer res.Body.Close()
-	r := ioutil.NopCloser(res.Body)
-	untarErr := Untar(r, dest)
-	if untarErr != nil {
-		return untarErr
-	}
-
-	return nil
 }
